fix: align code link and prefer package docs URL
- Add icon to code link for consistent alignment with other links
- Docs link now uses package homepage (typically their docs) when available
- Falls back to our generated API docs when no homepage exists
- Remove redundant homepage link (now covered by docs)
