+++ app/pages/[...package].vue
@@ -217,6 +217,30 @@ const homepageUrl = computed(() => {
   return homepage
 })
 
+// Docs URL: prefer package's homepage (often their docs site), fall back to our API docs
+const docsLink = computed(() => {
+  const homepage = displayVersion.value?.homepage
+  if (homepage) {
+    return {
+      href: homepage,
+      isExternal: true,
+    }
+  }
+
+  // Fall back to our generated API docs
+  if (displayVersion.value) {
+    return {
+      to: {
+        name: 'docs' as const,
+        params: { path: [...pkg.value!.name.split('/'), 'v', displayVersion.value.version] },
+      },
+      isExternal: false,
+    }
+  }
+
+  return null
+})
+
 function normalizeGitUrl(url: string): string {
   return url
     .replace(/^git\+/, '')
@@ -610,17 +634,7 @@ defineOgImageComponent('Package', {
                 {{ formatCompactNumber(stars, { decimals: 1 }) }}
               </a>
             </li>
-            <li v-if="homepageUrl">
-              <a
-                :href="homepageUrl"
-                target="_blank"
-                rel="noopener noreferrer"
-                class="link-subtle font-mono text-sm inline-flex items-center gap-1.5"
-              >
-                <span class="i-carbon-link w-4 h-4" aria-hidden="true" />
-                homepage
-              </a>
-            </li>
+
             <li v-if="displayVersion?.bugs?.url">
               <a
                 :href="displayVersion.bugs.url"
@@ -672,12 +686,20 @@ defineOgImageComponent('Package', {
               </a>
             </li>
 
-            <li v-if="displayVersion">
+            <li v-if="docsLink">
+              <a
+                v-if="docsLink.isExternal"
+                :href="docsLink.href"
+                target="_blank"
+                rel="noopener noreferrer"
+                class="link-subtle font-mono text-sm inline-flex items-center gap-1.5"
+              >
+                <span class="i-carbon-document w-4 h-4" aria-hidden="true" />
+                docs
+              </a>
               <NuxtLink
-                :to="{
-                  name: 'docs',
-                  params: { path: [...pkg.name.split('/'), 'v', displayVersion.version] },
-                }"
+                v-else
+                :to="docsLink.to"
                 class="link-subtle font-mono text-sm inline-flex items-center gap-1.5"
               >
                 <span class="i-carbon-document w-4 h-4" aria-hidden="true" />
@@ -693,6 +715,7 @@ defineOgImageComponent('Package', {
                 class="link-subtle font-mono text-sm inline-flex items-center gap-1.5"
                 aria-keyshortcuts="."
               >
+                <span class="i-carbon-code w-4 h-4" aria-hidden="true" />
                 code
                 <kbd
                   class="hidden sm:inline-flex items-center justify-center w-4 h-4 text-xs bg-bg-muted border border-border rounded"
